LU-3963 cleanup: C89 and build cleanups
use C99 initialization for a types, remove some platform depended code
and use POSIX compatible instead.
don't build a lhsmtool is none utils allowed.

Reviewed-on: http://review.whamcloud.com/7803
Tested-by: Jenkins
Tested-by: Maloo <hpdd-maloo@intel.com>

+++ libcfs/libcfs/posix/posix-debug.c
@@ -210,7 +210,7 @@ libcfs_debug_vmsg2(struct libcfs_debug_msg_data *msgdata,
         gettimeofday(&tv, NULL);
 
         fprintf(debug_file_fd, CFS_TIME_T".%06lu:%u:%s:(%s:%d:%s()): %s",
-                tv.tv_sec, tv.tv_usec, source_pid, source_nid,
+		tv.tv_sec, (unsigned long)tv.tv_usec, source_pid, source_nid,
                 msgdata->msg_file, msgdata->msg_line, msgdata->msg_fn, buf);
 
         return 0;
